# api/public/swagger.yaml
swagger: '2.0'
info:
  version: "0.0.1"
  title: ESG API 
consumes:
  - text/plain
produces:
  - application/json
paths:
...
  /info/{symbol}:
    get:
      operationId: getInfo
      description: Get details for a given company 
      tags:
        - stocks 
      parameters:
        - in: path 
          name: symbol 
          schema:
            type: string 
          required: true
          description: Company's stock symbol 
      responses:
        200:
          description: "Company details"
          examples:
            application/json: {"cfc_company": "MSFT", "info": { "zip": "98052-6399", "sector": "Technology", "fullTimeEmployees": 163000, "longBusinessSummary": "Microsoft Corporation develops, licenses, and supports software, services, devices, and solutions worldwide. Its Productivity and Business Processes segment offers Office, Exchange, SharePoint, Microsoft Teams, Office 365 Security and Compliance, and Skype for Business, as well as related Client Access Licenses (CAL); Skype, Outlook.com, OneDrive, and LinkedIn; and Dynamics 365, a set of cloud-based and on-premises business solutions for small and medium businesses, organizations, and enterprise divisions. Its Intelligent Cloud segment licenses SQL and Windows Servers, Visual Studio, System Center, and related CALs; GitHub that provides a collaboration platform and code hosting service for developers; and Azure, a cloud platform. It also offers support services and Microsoft consulting services to assist customers in developing, deploying, and managing Microsoft server and desktop solutions; and training and certification to developers and IT professionals on Microsoft products. Its More Personal Computing segment provides Windows original equipment manufacturer (OEM) licensing and other non-volume licensing of the Windows operating system; Windows Commercial, such as volume licensing of the Windows operating system, Windows cloud services, and other Windows commercial offerings; patent licensing; Windows Internet of Things; and MSN advertising. It also offers Surface, PC accessories, PCs, tablets, gaming and entertainment consoles, and other devices; Gaming, including Xbox hardware, and Xbox content and services; video games and third-party video game royalties; and Search, including Bing and Microsoft advertising. It sells its products through OEMs, distributors, and resellers; and directly through digital marketplaces, online stores, and retail stores. It has a collaboration with DXC Technology, Dynatrace, Inc., Morgan Stanley, Micro Focus, WPP plc, and iCIMS, Inc. The company was founded in 1975 and is headquartered in Redmond, Washington.", "city": "Redmond", "phone": "425 882 8080", "state": "WA", "country": "United States", "companyOfficers": [], "website": "http://www.microsoft.com", "maxAge": 1, "address1": "One Microsoft Way", "industry": "Software—Infrastructure", "ebitdaMargins": 0.47245, "profitMargins": 0.35016, "grossMargins": 0.68381, "operatingCashflow": 72703000576, "revenueGrowth": 0.191, "operatingMargins": 0.40150002, "ebitda": 75577999360, "targetLowPrice": 256.7, "recommendationKey": "buy", "grossProfits": 96937000000, "freeCashflow": 37776875520, "targetMedianPrice": 300, "currentPrice": 286.14, "earningsGrowth": 0.45, "currentRatio": 2.294, "returnOnAssets": 0.13508001, "numberOfAnalystOpinions": 33, "targetMeanPrice": 298.92, "debtToEquity": 60.414, "returnOnEquity": 0.44991, "targetHighPrice": 340, "totalCash": 125013000192, "totalDebt": 81260003328, "totalRevenue": 159969001472, "totalCashPerShare": 16.599, "financialCurrency": "USD", "revenuePerShare": 21.156, "quickRatio": 2.096, "recommendationMean": 1.6, "exchange": "NMS", "shortName": "Microsoft Corporation", "longName": "Microsoft Corporation", "exchangeTimezoneName": "America/New_York", "exchangeTimezoneShortName": "EDT", "isEsgPopulated": true, "gmtOffSetMilliseconds": "-14400000", "underlyingSymbol": null, "quoteType": "EQUITY", "symbol": "MSFT", "underlyingExchangeSymbol": null, "headSymbol": null, "messageBoardId": "finmb_21835", "uuid": "b004b3ec-de24-385e-b2c1-923f10d3fb62", "market": "us_market", "annualHoldingsTurnover": null, "enterpriseToRevenue": 12.799, "beta3Year": null, "enterpriseToEbitda": 27.091, "52WeekChange": null, "morningStarRiskRating": null, "forwardEps": 8.36, "revenueQuarterlyGrowth": null, "sharesOutstanding": 7531570176, "fundInceptionDate": null, "annualReportExpenseRatio": null, "totalAssets": null, "bookValue": 17.853, "sharesShort": 48137624, "sharesPercentSharesOut": 0.0064, "fundFamily": null, "lastFiscalYearEnd": 1593475200, "heldPercentInstitutions": 0.72033, "netIncomeToCommon": 56014999552, "trailingEps": 7.338, "lastDividendValue": null, "SandP52WeekChange": null, "priceToBook": 16.02756, "heldPercentInsiders": 0.00071999995, "nextFiscalYearEnd": 1656547200, "yield": null, "mostRecentQuarter": 1617148800, "shortRatio": 2.24, "sharesShortPreviousMonthDate": 1620950400, "floatShares": 7420483826, "beta": 0.787812, "enterpriseValue": 2047464046592, "priceHint": 2, "threeYearAverageReturn": null, "lastSplitDate": 1045526400, "lastSplitFactor": "2:1", "legalType": null, "morningStarOverallRating": null, "earningsQuarterlyGrowth": 0.438, "priceToSalesTrailing12Months": 13.471883, "dateShortInterest": 1623715200, "pegRatio": 1.95, "ytdReturn": null, "forwardPE": 34.227276, "lastCapGain": null, "shortPercentOfFloat": 0.0064, "sharesShortPriorMonth": 40526093, "category": null, "fiveYearAverageReturn": null, "previousClose": 281.4, "regularMarketOpen": 283.84, "twoHundredDayAverage": 246.87914, "trailingAnnualDividendYield": 0.0077825165, "payoutRatio": 0.2916, "volume24Hr": null, "regularMarketDayHigh": 286.42, "navPrice": null, "averageDailyVolume10Day": 25062350, "regularMarketPreviousClose": 281.4, "fiftyDayAverage": 267.24344, "trailingAnnualDividendRate": 2.19, "open": 283.84, "averageVolume10days": 25062350, "expireDate": null, "algorithm": null, "dividendRate": 2.24, "exDividendDate": 1629244800, "circulatingSupply": null, "startDate": null, "regularMarketDayLow": 283.42, "currency": "USD", "trailingPE": 38.994278, "regularMarketVolume": 21770109, "lastMarket": null, "maxSupply": null, "openInterest": null, "marketCap": 2155083595776, "volumeAllCurrencies": null, "strikePrice": null, "averageVolume": 24872577, "dayLow": 283.42, "ask": 286.21, "askSize": 3200, "volume": 21770109, "fiftyTwoWeekHigh": 286.42, "fromCurrency": null, "fiveYearAvgDividendYield": 1.57, "fiftyTwoWeekLow": 196.25, "bid": 286.2, "tradeable": false, "dividendYield": 0.0081, "bidSize": 1200, "dayHigh": 286.42, "regularMarketPrice": 286.14, "logo_url": "https://logo.clearbit.com/microsoft.com" }, "market_data": { "Open": { "2021-06-23T00:00:00.000Z": 265.9899902344, "2021-06-24T00:00:00.000Z": 266.1600036621, "2021-06-25T00:00:00.000Z": 266.2300109863, "2021-06-28T00:00:00.000Z": 266.1900024414, "2021-06-29T00:00:00.000Z": 268.8699951172, "2021-06-30T00:00:00.000Z": 270.6900024414, "2021-07-01T00:00:00.000Z": 269.6099853516, "2021-07-02T00:00:00.000Z": 272.8200073242, "2021-07-06T00:00:00.000Z": 278.0299987793, "2021-07-07T00:00:00.000Z": 279.3999938965, "2021-07-08T00:00:00.000Z": 276.8999938965, "2021-07-09T00:00:00.000Z": 275.7200012207, "2021-07-12T00:00:00.000Z": 279.1600036621, "2021-07-13T00:00:00.000Z": 277.5199890137, "2021-07-14T00:00:00.000Z": 282.3500061035, "2021-07-15T00:00:00.000Z": 282, "2021-07-16T00:00:00.000Z": 282.0700073242, "2021-07-19T00:00:00.000Z": 278.9299926758, "2021-07-20T00:00:00.000Z": 278.0299987793, "2021-07-21T00:00:00.000Z": 278.8999938965, "2021-07-22T00:00:00.000Z": 283.8399963379 }, "High": { "2021-06-23T00:00:00.000Z": 266.8299865723, "2021-06-24T00:00:00.000Z": 267.8500061035, "2021-06-25T00:00:00.000Z": 267.25, "2021-06-28T00:00:00.000Z": 268.8999938965, "2021-06-29T00:00:00.000Z": 271.6499938965, "2021-06-30T00:00:00.000Z": 271.3599853516, "2021-07-01T00:00:00.000Z": 271.8399963379, "2021-07-02T00:00:00.000Z": 278, "2021-07-06T00:00:00.000Z": 279.3699951172, "2021-07-07T00:00:00.000Z": 280.6900024414, "2021-07-08T00:00:00.000Z": 278.7300109863, "2021-07-09T00:00:00.000Z": 278.049987793, "2021-07-12T00:00:00.000Z": 279.7699890137, "2021-07-13T00:00:00.000Z": 282.8500061035, "2021-07-14T00:00:00.000Z": 283.6600036621, "2021-07-15T00:00:00.000Z": 282.5100097656, "2021-07-16T00:00:00.000Z": 284.1000061035, "2021-07-19T00:00:00.000Z": 280.3699951172, "2021-07-20T00:00:00.000Z": 280.9700012207, "2021-07-21T00:00:00.000Z": 281.5199890137, "2021-07-22T00:00:00.000Z": 286.4200134277 }, "Low": { "2021-06-23T00:00:00.000Z": 264.4299926758, "2021-06-24T00:00:00.000Z": 265.4700012207, "2021-06-25T00:00:00.000Z": 264.7600097656, "2021-06-28T00:00:00.000Z": 265.9100036621, "2021-06-29T00:00:00.000Z": 267.9800109863, "2021-06-30T00:00:00.000Z": 269.6000061035, "2021-07-01T00:00:00.000Z": 269.6000061035, "2021-07-02T00:00:00.000Z": 272.5, "2021-07-06T00:00:00.000Z": 274.299987793, "2021-07-07T00:00:00.000Z": 277.1499938965, "2021-07-08T00:00:00.000Z": 274.8699951172, "2021-07-09T00:00:00.000Z": 275.3200073242, "2021-07-12T00:00:00.000Z": 276.5799865723, "2021-07-13T00:00:00.000Z": 277.3900146484, "2021-07-14T00:00:00.000Z": 280.549987793, "2021-07-15T00:00:00.000Z": 279.8299865723, "2021-07-16T00:00:00.000Z": 279.4599914551, "2021-07-19T00:00:00.000Z": 274.450012207, "2021-07-20T00:00:00.000Z": 276.2600097656, "2021-07-21T00:00:00.000Z": 277.2900085449, "2021-07-22T00:00:00.000Z": 283.4200134277 }, "Close": { "2021-06-23T00:00:00.000Z": 265.2699890137, "2021-06-24T00:00:00.000Z": 266.6900024414, "2021-06-25T00:00:00.000Z": 265.0199890137, "2021-06-28T00:00:00.000Z": 268.7200012207, "2021-06-29T00:00:00.000Z": 271.3999938965, "2021-06-30T00:00:00.000Z": 270.8999938965, "2021-07-01T00:00:00.000Z": 271.6000061035, "2021-07-02T00:00:00.000Z": 277.6499938965, "2021-07-06T00:00:00.000Z": 277.6600036621, "2021-07-07T00:00:00.000Z": 279.9299926758, "2021-07-08T00:00:00.000Z": 277.4200134277, "2021-07-09T00:00:00.000Z": 277.9400024414, "2021-07-12T00:00:00.000Z": 277.3200073242, "2021-07-13T00:00:00.000Z": 280.9800109863, "2021-07-14T00:00:00.000Z": 282.5100097656, "2021-07-15T00:00:00.000Z": 281.0299987793, "2021-07-16T00:00:00.000Z": 280.75, "2021-07-19T00:00:00.000Z": 277.0100097656, "2021-07-20T00:00:00.000Z": 279.3200073242, "2021-07-21T00:00:00.000Z": 281.3999938965, "2021-07-22T00:00:00.000Z": 286.1400146484 }, "Volume": { "2021-06-23T00:00:00.000Z": 19518700, "2021-06-24T00:00:00.000Z": 21446900, "2021-06-25T00:00:00.000Z": 25611100, "2021-06-28T00:00:00.000Z": 19590000, "2021-06-29T00:00:00.000Z": 19937800, "2021-06-30T00:00:00.000Z": 21656500, "2021-07-01T00:00:00.000Z": 16725300, "2021-07-02T00:00:00.000Z": 26458000, "2021-07-06T00:00:00.000Z": 31565600, "2021-07-07T00:00:00.000Z": 23260000, "2021-07-08T00:00:00.000Z": 24618600, "2021-07-09T00:00:00.000Z": 23916700, "2021-07-12T00:00:00.000Z": 18931700, "2021-07-13T00:00:00.000Z": 26120100, "2021-07-14T00:00:00.000Z": 23113700, "2021-07-15T00:00:00.000Z": 22604200, "2021-07-16T00:00:00.000Z": 26186800, "2021-07-19T00:00:00.000Z": 32935600, "2021-07-20T00:00:00.000Z": 26259700, "2021-07-21T00:00:00.000Z": 24357700, "2021-07-22T00:00:00.000Z": 21770109 }, "Dividends": { "2021-06-23T00:00:00.000Z": 0, "2021-06-24T00:00:00.000Z": 0, "2021-06-25T00:00:00.000Z": 0, "2021-06-28T00:00:00.000Z": 0, "2021-06-29T00:00:00.000Z": 0, "2021-06-30T00:00:00.000Z": 0, "2021-07-01T00:00:00.000Z": 0, "2021-07-02T00:00:00.000Z": 0, "2021-07-06T00:00:00.000Z": 0, "2021-07-07T00:00:00.000Z": 0, "2021-07-08T00:00:00.000Z": 0, "2021-07-09T00:00:00.000Z": 0, "2021-07-12T00:00:00.000Z": 0, "2021-07-13T00:00:00.000Z": 0, "2021-07-14T00:00:00.000Z": 0, "2021-07-15T00:00:00.000Z": 0, "2021-07-16T00:00:00.000Z": 0, "2021-07-19T00:00:00.000Z": 0, "2021-07-20T00:00:00.000Z": 0, "2021-07-21T00:00:00.000Z": 0, "2021-07-22T00:00:00.000Z": 0 }, "Stock Splits": { "2021-06-23T00:00:00.000Z": 0, "2021-06-24T00:00:00.000Z": 0, "2021-06-25T00:00:00.000Z": 0, "2021-06-28T00:00:00.000Z": 0, "2021-06-29T00:00:00.000Z": 0, "2021-06-30T00:00:00.000Z": 0, "2021-07-01T00:00:00.000Z": 0, "2021-07-02T00:00:00.000Z": 0, "2021-07-06T00:00:00.000Z": 0, "2021-07-07T00:00:00.000Z": 0, "2021-07-08T00:00:00.000Z": 0, "2021-07-09T00:00:00.000Z": 0, "2021-07-12T00:00:00.000Z": 0, "2021-07-13T00:00:00.000Z": 0, "2021-07-14T00:00:00.000Z": 0, "2021-07-15T00:00:00.000Z": 0, "2021-07-16T00:00:00.000Z": 0, "2021-07-19T00:00:00.000Z": 0, "2021-07-20T00:00:00.000Z": 0, "2021-07-21T00:00:00.000Z": 0, "2021-07-22T00:00:00.000Z": 0 } }} 
...